contribution: add new vitesse community variation
This pull request proposes to reference my own variation of `vitesse` and `vitesse-stackter-clean-architect` and focuses on adding strong out of the box linting and formatting.

Code style is verified on pre-commit. Component and E2E tests run on-pre-push.

As a bonus, the pull request also adds bug, user story and pull request templates for github. These are also added for gitlab with a gitlab CI/CD script.